requirements.txt
faiss-cpu==1.8.0
google-generativeai==0.5.3
langchain==0.2.5
langchain-community==0.0.36
opencv-python==4.9.0.80
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
tqdm==4.66.4
streamlit_webrtc==0.47.7